Address review: move get_capacity_unit_label to forms/utils, unit-agnostic verbose_names, model form label overrides
